requirements.txt
streamlit
transformers
torch
torchvision
pillow
pyttsx3
groq
python-dotenv
accelerate
bitsandbytes
peft